datos sincronizados

+++ net/ATL.Servicio/ATL.Entidad/Talla.cs
@@ -8,8 +8,8 @@ namespace ATL.Entidad
     public class Talla
     {
         public long codigoProducto { get; set; }
-        public int numeroTalla { get; set; }
-        public int stockDisponibleTalla { get; set; }
+        public long numeroTalla { get; set; }
+        public long stockDisponibleTalla { get; set; }
 
     }
 }
